glChartcanvas: make ocpn_gl_options.h a separate file
Makes it possible to reference ocpnGLOptions also in non-GL builds,
decreasing the need for evil ifdefs.
