+++ gui/include/gui/ocpn_gl_options.h
@@ -1,5 +1,5 @@
-/**************************************************************************
- *   Copyright (C) 2025 Alec Leamas                                        *
+/***************************************************************************
+ *   Copyright (C) 2013 by David S. Register                               *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
  *   it under the terms of the GNU General Public License as published by  *
@@ -17,31 +17,34 @@
 
 /**
  * \file
- * Platform independent GL includes
+ * Class ocpnGLOptions -- OpenGL runtime options
  */
 
-#if defined(__ANDROID__)
-#include <qopengl.h>
-#include <GL/gl_private.h>  // this is a cut-down version of gl.h
-#include <GLES2/gl2.h>
+class ocpnGLOptions {
+public:
+  bool m_bUseAcceleratedPanning;
 
-#elif defined(ocpnUSE_GL)
-#if defined(_WIN32)
-#include <glew.h>
+  /**
+   * Controls OpenGL canvas hardware-accelerated panning mode.
+   *
+   * When true, uses an OpenGL optimization where the chart is rendered to a
+   * texture (via FBO - Frame Buffer Object) during panning operations.
+   * This texture is then translated/moved using GPU operations rather than
+   * redrawing the entire chart for each pan event, significantly improving
+   * performance during chart panning.
+   *
+   * This is separate from standard chart panning which is enabled by default
+   * and uses left-click + drag. This flag only controls whether that panning
+   * uses hardware acceleration.
+   */
+  bool m_bUseCanvasPanning;
 
-#elif __APPLE__
-typedef void (*_GLUfuncptr)();
-#define GL_COMPRESSED_RGB_FXT1_3DFX 0x86B0
-#include <OpenGL/gl.h>
-#include <OpenGL/glu.h>
+  bool m_bTextureCompression;
+  bool m_bTextureCompressionCaching;
 
-#elif defined(__linux__)
-#include <GL/glew.h>
-#include <GL/gl.h>
-#include <GL/glu.h>
-#include <GL/glext.h>
+  int m_iTextureDimension;
+  int m_iTextureMemorySize;
 
-#else
-#error platform not supported.
-#endif  // _win32
-#endif  // ocpnUSE_GL
+  bool m_GLPolygonSmoothing;
+  bool m_GLLineSmoothing;
+};
